chore(card): move test images to tests dir
