+++ src/card/__tests__/card-image-object.scenario.tsx
@@ -7,7 +7,7 @@ LICENSE file in the root directory of this source tree.
 import * as React from 'react';
 
 import { Card, type ImageProps, StyledBody } from '..';
-import { header as headerImg, thumbnail as thumbnailImg } from '../images';
+import { header as headerImg, thumbnail as thumbnailImg } from './images';
 import { styled } from '../../styles';
 
 const Container = styled('div', { width: '328px' });
