--- src/card/index.js.flow
@@ -1,21 +1,0 @@
-/*
-Copyright (c) Uber Technologies, Inc.
-
-This source code is licensed under the MIT license found in the
-LICENSE file in the root directory of this source tree.
-*/
-
-/* @flow */
-
-export { default as Card, hasThumbnail } from './card.js';
-export {
-  Action as StyledAction,
-  Body as StyledBody,
-  Contents as StyledContents,
-  HeaderImage as StyledHeaderImage,
-  Thumbnail as StyledThumbnail,
-  Title as StyledTitle,
-  Root as StyledRoot,
-  Root as StyledWrapper,
-} from './styled-components.js';
-export type * from './types.js';
